Moved calc_piecewisefit to calibration.py
