+++ pyowc/__init__.py
@@ -19,7 +19,6 @@ from .plot import dashboard
 from . import calibration
 from . import configuration
 
-
 #
 __all__ = (
     # Sub-packages,
